creates a separate component for input fields
